+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## Unreleased
+
+### 🐞 Fixes
+
+* **exchanges:** refresh widget logos from CoinGecko manifest
+
 ## 1.0.0 (2025-10-10)